Optimize evaluation memory: use float16 for logits storage
- Store Mode 1 logits as float16 instead of float32
- Memory usage: 31GB -> 15.5GB (50% reduction)
- Convert back to float32 for cross_entropy computation
- No accuracy loss: float16 for storage, float32 for computation

Fixes OOM issue during 5-mode evaluation with max_eval_batches=10

Modified:
- train/evaluation_modes.py:97: Add .half() to logits storage
- train/evaluation_modes.py:408: Add .float() for Mode 4 computation
- train/evaluation_modes.py:470: Add .float() for Mode 5 computation

+++ train/evaluation_modes.py
@@ -93,7 +93,8 @@ def evaluate_mode1_autoregressive(model, dataloader, device, max_batches=None):
             num_batches += 1
 
             # Store for Mode 4/5 (move to CPU and free GPU memory immediately)
-            all_logits.append(logits.cpu())
+            # Convert to float16 to save memory (31GB → 15.5GB)
+            all_logits.append(logits.cpu().half())
             all_labels.append(input_ids.cpu())
             del logits, input_ids  # Free GPU memory
             torch.cuda.empty_cache()  # Clear cache
@@ -404,7 +405,7 @@ def evaluate_mode4_cross_boundary(logits_list, labels_list, eval_indices_list):
 
             if valid_mask.sum() > 0:
                 loss = F.cross_entropy(
-                    shift_logits[valid_mask],
+                    shift_logits[valid_mask].float(),  # Convert float16 back to float32 for computation
                     shift_labels[valid_mask],
                     reduction='sum'
                 )
@@ -466,7 +467,7 @@ def evaluate_mode5_cross_training(logits_list, labels_list, eval_indices_list):
 
             if valid_mask.sum() > 0:
                 loss = F.cross_entropy(
-                    shift_logits[valid_mask],
+                    shift_logits[valid_mask].float(),  # Convert float16 back to float32 for computation
                     shift_labels[valid_mask],
                     reduction='sum'
                 )
